Revert back to returning extensions in PluginInstance.
This has been a not so well though out design decision. Plugins should not be
able to (de)register extensions at any time. 1. This is a nightmare to get
threadsafe. We want plugins to load in background. Soon. 2. UX suffers, instead
of being able to toggle extensions rather add another plugin.