API: Change the plugin loading design
- Revert to threaded plugin loading. See rationale below.
- Move frontend loading into plugin registry

API:

- `albert::Frontend`
  - Add function `QString id() const`
- `albert::PluginLoader`
  - Rem function `load()`
  - Rem function `PluginInstance *createInstance()`
  - Add function `PluginInstance &load()`

Some facts on parallelization:

- Parallelized loading or instantiation of Python plugins is not possible
  since the GIL serializes any Python execution.
- Parallelized loading of native plugins is not possible, because either
  QPluginloader or the platform linker serialize the threads.
- Parallelized instantiation of native plugins may be possible, but:
  - QPluginLoader instantiates the plugin on load, i.e., as mentioned above,
    it will effectively be serialized. This could be worked around by an
    indirection using a PluginInstance factory method though.
  - For the sake of clean interfaces static injection of the plugin metadata
    is used, which requires serialized instantiation. This could be redesigned
    such that it uses setter injection.
  - The parts of the public API required by the plugins while instantiation
    are not thread safe. Obviously the solution is to make them thread safe
    which is lots of work.

Conclusion:

The only benefit we could get (with lots of work) is the parallel construction
of the native plugins. Most of them dont actually do something in the
constructor (well besides the Python plugin). So it's not worth the hassle.
The real resource hog is the dynamic loading of the library which cannot be
parallelized effectively. So actually (currently) theres no point in multi-
threading plugin loading and initialization at all.

However:

There is still the idea of putting the library loading process into a
background thread to get a more responsive UI. This is feasable, but the
benefits are again marginal. Improvements here would be recognizable only
at start up and upon clicking a checkbox in the plugins tab. Maybe in future
but currently there is more important stuff to do.